twin: catalog-level editorial fields, parallel sims, sim_progress streaming (D-16/D-17/D-18)
Propose layer now ranks candidates by catalog signals first, then by
twin-computed risk. Three orthogonal editorial fields join the runtime
risk_score so the frontend can render a sortable candidates table
without re-reading the registry:

  - effort_score (float 0..1)
  - mission_impact (none | minor | major | mission-ending)
  - reversibility (trivial | easy | hard)

Propose pre-filters to top-K=5 by a lexicographic coarse key
(effort, impact, reversibility) before paying the twin simulation
cost, then runs those 5 in a ThreadPoolExecutor (max_workers=5).
Each worker calls validate_procedure, which fires an on_step
callback after every step; the bridge forwards that callback to a
queue.Queue in AppState, which the WS producer drains and attaches
to ticks as sim_progress (at most one event per tick, drop-oldest
overflow at 2000). The WS contract now carries the same Proposal
shape the Propose layer returns, with candidates_ranked: List[RankedCandidate]
carrying all 5 fields per row.

The BIBLE picks up D-16 (catalog fields), D-17 (parallel sims +
progress streaming), D-18 (RankedCandidate replaces the tuple), and
a §3.2 Proposal/RankedCandidate TypeScript type that matches the
landed wire shape. Stage 3, Stage 4, the bridge section, the
test gate, and the procedure table are all updated to match.

Tests: 42 twin tests pass, 43 legacy tests pass, 1 skipped.

+++ live/error_stream.py
@@ -220,7 +220,7 @@ class ErrorWindow:
         E_seq_max = np.array([max(e_s[e[0] : e[1] + 1]) for e in E_seq])
         E_seq_max_sorted = np.sort(E_seq_max)[::-1]
         E_seq_max_sorted = np.append(E_seq_max_sorted, [non_anom_max])
-        i_to_remove = np.array([])
+        i_to_remove = np.array([], dtype=int)
         for i in range(0, len(E_seq_max_sorted) - 1):
             if (
                 E_seq_max_sorted[i] - E_seq_max_sorted[i + 1]
@@ -230,7 +230,7 @@ class ErrorWindow:
                     np.argwhere(E_seq_max == E_seq_max_sorted[i]),
                 )
             else:
-                i_to_remove = np.array([])
+                i_to_remove = np.array([], dtype=int)
         i_to_remove[::-1].sort()
         if len(i_to_remove) > 0:
             E_seq = np.delete(E_seq, i_to_remove, axis=0)
